requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-core>=0.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
duckdb>=0.9.0
langchain-community>=0.0.1
